add test for gradient descent
